WaveGreen2D

Documentation for WaveGreen2D.

WaveGreen2D.NearField.ReducedChebyshevSeries

WaveGreen2D.WaveParameters

WaveGreen2D.NearField.Gᴺ

WaveGreen2D.NearField.setintegrals!

WaveGreen2D.find_k₀

WaveGreen2D.find_kₙ

WaveGreen2D.findroot

WaveGreen2D.setwave!

WaveGreen2D.WaveParameters — Type

WaveParameters

Dimensional parameters that define the environmental conditions. The wavenumber and evanescent wavenumbers must be computed from the others via the function setwave!.

Fields

depth::Float64: water depth (m)

frequency::Float64: wave frequency (rad/s)

gravity::Float64: acceleration of gravity (m/s²)

wavenumber::Float64: wavenumber (1/m⁻¹)

evanumbers::NTuple{n, Float64}: n evanescent wavenumbers (1/m⁻¹)

WaveGreen2D.find_k₀ — Function

find_k₀(h::Real, ω::Real, g::Real=9.80665) -> Float64

Finds the wavenumber k₀ from the dispersion relation $ω^2 = k g \tanh(k h)$.

WaveGreen2D.find_kₙ — Function

find_kₙ(n::Int, h::Real, ω::Real, g::Real=9.80665) -> Float64

Finds the evanescent wavenumber kₙ from the dispersion relation $ω^2 = -k g \tan(k h)$.

WaveGreen2D.findroot — Function

findroot(y, y′, x₀::Real, tol::Real=eps(), nmax::Int=50) -> Float64

Finds the root of y closer to x₀, if this root exists, using the Newton-Raphson method.

Arguments

y: the function to find the root of

y′: the derivative of y

x₀::Real: initial estimate for the root of y

tol::Real=eps(): tolerance for the absolute error

nmax::Int=50: maximum number of iterations

Returns

Float64: final estimate for the root of y

Warnings

A warning is raised if the maximum number of iterations is reached without convergence.

WaveGreen2D.setwave! — Method

function setwave!(depth::Real, frequency::Real, gravity::Real=9.80665) -> Nothing

Sets the parameters that define the environmental conditions.

WaveGreen2D.NearField.ReducedChebyshevSeries — Type

ReducedChebyshevSeries

Chebyshev series approximations of L₁ and L₂ evaluated for a fixed value of $H = h ω^2 / g$.

WaveGreen2D.NearField.Gᴺ — Method

Gᴺ(
    field_point::SVector{2,Float64}, source_point::SVector{2,Float64}
) -> Float64, SVector{2,Float64}

Finite-depth free surface Green function for field and source points close to each other, which is defined by the dimensionless horizontal distance A ≤ 0.5.

WaveGreen2D.NearField.setintegrals! — Method

setintegrals!(H::Float64) -> Nothing

Sets the Chebyshev series approximations of $L₁$ and $L₂$ for a fixed parameter $H = h ω^2 / g$. Shallow and intermediate waters are defined by $H ≤ π$, while $H = 0.01$ is the minimum value that could be used to compute the analytical expressions of $L₁$ and $L₂$. The range $π < H ≤ 7$ is used solely for comparison with the deep water formulation and not used to compute the near field Green function.