WaveGreen2D

Documentation for WaveGreen2D.

WaveGreen2D.Chebyshev.AbstractChebyshevSeries

WaveGreen2D.Chebyshev.AbstractChebyshevSeries

WaveGreen2D.Chebyshev.ChebyshevCluster

WaveGreen2D.Chebyshev.ChebyshevCluster

WaveGreen2D.Chebyshev.ChebyshevCluster

WaveGreen2D.Chebyshev.ChebyshevSeries

WaveGreen2D.Chebyshev.ChebyshevSeries

WaveGreen2D.Chebyshev.ChebyshevSeries

WaveGreen2D.Chebyshev.ChebyshevSeries

WaveGreen2D.Chebyshev.TransformedChebyshevSeries

WaveGreen2D.Chebyshev.TransformedChebyshevSeries

WaveGreen2D.Chebyshev.TransformedChebyshevSeries

WaveGreen2D.Chebyshev.clenshaw

WaveGreen2D.Chebyshev.clenshaw

WaveGreen2D.Chebyshev.contains

WaveGreen2D.Chebyshev.contains

WaveGreen2D.Chebyshev.contains

WaveGreen2D.Chebyshev.domain

WaveGreen2D.Chebyshev.gradient

WaveGreen2D.Chebyshev.gradient

WaveGreen2D.Chebyshev.gradient

WaveGreen2D.Chebyshev.gradient

WaveGreen2D.Chebyshev.gradient

WaveGreen2D.Chebyshev.gradient_clenshaw

WaveGreen2D.Chebyshev.gradient_clenshaw

WaveGreen2D.Chebyshev.hessian

WaveGreen2D.Chebyshev.hessian

WaveGreen2D.Chebyshev.hessian

WaveGreen2D.Chebyshev.hessian

WaveGreen2D.Chebyshev.hessian

WaveGreen2D.Chebyshev.hessian_clenshaw

WaveGreen2D.Chebyshev.hessian_clenshaw

WaveGreen2D.Chebyshev.normalize

WaveGreen2D.Chebyshev.order

WaveGreen2D.Chebyshev.order

WaveGreen2D.Chebyshev.symmatrix

WaveGreen2D.Chebyshev.validate_transformation

WaveGreen2D.Chebyshev.AbstractChebyshevSeries — Method

(f::AbstractChebyshevSeries{T, 1})(x::T) where T -> T

Simpler callable constructor for evaluating a one-dimensional Chebyshev series f at a point x, where x is of type T.

WaveGreen2D.Chebyshev.AbstractChebyshevSeries — Method

(f::AbstractChebyshevSeries{T, N})(x::AbstractVector{T}) where {T, N} -> T

Simpler callable constructor for evaluating a Chebyshev series f at a point x, where x is of any subtype of an AbstractVector{T}.

WaveGreen2D.Chebyshev.ChebyshevCluster — Type

ChebyshevCluster{T, N, M}

A collection of M equidimensional ChebyshevSeries objects.

Fields

series::NTuple{M, ChebyshevSeries{T, N}}: N-dimensional Chebyshev series

WaveGreen2D.Chebyshev.ChebyshevCluster — Method

ChebyshevCluster(
    f¹::ChebyshevSeries{T,N}, fᴹ⁻¹::ChebyshevSeries{T,N}...
) where {T,N} -> ChebyshevCluster{T,N,M}

Simpler constructor for a Chebyshev cluster.

WaveGreen2D.Chebyshev.ChebyshevCluster — Method

(h::ChebyshevCluster{T, N})(x::SVector{N, T}) where {T, N} -> T

Evaluates the Chebyshev cluster h at a point x.

WaveGreen2D.Chebyshev.ChebyshevSeries — Type

ChebyshevSeries{T, N}

The Chebyshev series approximation of a N-dimensional function defined in a bounded domain.

Fields

coefs::Array{T, N}: coefficients

lb::SVector{N, T}: domain lower bound

ub::SVector{N, T}: domain upper bound

WaveGreen2D.Chebyshev.ChebyshevSeries — Method

(f::ChebyshevSeries{T, N})(x::SVector{N, T}) where {T, N} -> T

Evaluates the Chebyshev series f at a point x.

WaveGreen2D.Chebyshev.ChebyshevSeries — Method

ChebyshevSeries(coefs::Array{T, 0}, lb::SVector{0, T}, ub::SVector{0, T}) where T -> T

A zero-dimensional Chebyshev series is just the coefficient.

WaveGreen2D.Chebyshev.ChebyshevSeries — Method

ChebyshevSeries(coefs::Array{T, 1}, lb::T, ub::T) where T -> ChebyshevSeries{T, 1}

Simpler constructor for one-dimensional Chebyshev series.

WaveGreen2D.Chebyshev.TransformedChebyshevSeries — Type

TransformedChebyshevSeries{T, N}

The transformation of a Chebyshev series in the domain u to the domain x.

Fields

series::ChebyshevSeries{T, N}: series in the domain u

u::F: transformation function $u(x)$

∇u::G: transformation function gradient $∇u(x)$

Hu::H: transformation function hessian $\mathrm{H}u(x)$

WaveGreen2D.Chebyshev.TransformedChebyshevSeries — Method

(f::ChebyshevSeries{T, N})(x::SVector{N, T}) where {T, N} -> T

Evaluates the transformed Chebyshev series g at a point x, which is equivalent to evaluating g.series at the point g.u(x).

WaveGreen2D.Chebyshev.TransformedChebyshevSeries — Method

TransformedChebyshevSeries(
    series::ChebyshevSeries{T, N}
) where {T, N} -> TransformedChebyshevSeries{T, N}

Identically transformed Chebyshev series.

WaveGreen2D.Chebyshev.clenshaw — Method

clenshaw(a::Array{T, N}, x::SVector{N, T}) where {T, N} -> T

Implements the Clenshaw algorithm to evaluate the N-th dimensional Chebyshev series with coefficients a at a normalized point x in $[-1, 1]^N$.

WaveGreen2D.Chebyshev.clenshaw — Method

clenshaw(a::Array{T, N}, x::T) where {T, N} -> Array{T, N-1}

Implements the Clenshaw algorithm to evaluate the N-th dimensional Chebyshev series with coefficients a at a normalized value x of its N-th dimension.

WaveGreen2D.Chebyshev.contains — Method

contains(h::ChebyshevCluster{T, N, M}, x::SVector{N, T}) where {T, N, M} -> Int

Finds the index of the series in the cluster h where the point x is located. Returns 0 if x is not in h.

WaveGreen2D.Chebyshev.contains — Method

contains(f::ChebyshevSeries{T, N}, x::SVector{N, T}) where {T, N} -> Bool

Checks if the point x is in the domain of f.

WaveGreen2D.Chebyshev.contains — Method

contains(g::TransformedChebyshevSeries{T, N}, x::SVector{N, T}) where {T, N} -> Bool

Checks if the point x is in the domain of g, which is equivalent to checking if the point g.u(x) is in the domain of g.series.

WaveGreen2D.Chebyshev.domain — Method

domain(f::ChebyshevSeries{T,N}) where {T,N} -> String

Domain of the Chebyshev series given as the cartesian product of closed intervals along each dimension. The lower and upper bounds of the intervals are rounded to three digits of decimal precision.

WaveGreen2D.Chebyshev.gradient — Method

gradient(
    f::AbstractChebyshevSeries{T, N}, x::AbstractVector{T}
) where {T, N} -> T, SVector{N, T}

Simpler function for evaluating a Chebyshev series f and its gradient at a point x, where x is of any subtype of an AbstractVector{T}.

WaveGreen2D.Chebyshev.gradient — Method

gradient(h::ChebyshevCluster{T, N}, x::SVector{N, T}) where {T, N} -> T, SVector{N, T}

Evaluates the Chebyshev cluster h and its gradient at a point x.

WaveGreen2D.Chebyshev.gradient — Method

gradient(f::ChebyshevSeries{T, N}, x::SVector{N, T}) where {T, N} -> T, SVector{N, T}

Evaluates the Chebyshev series f and its gradient at a point x.

WaveGreen2D.Chebyshev.gradient — Method

gradient(
    g::TransformedChebyshevSeries{T, N}, x::SVector{N, T}
) where {T, N} -> T, SVector{N, T}

Evaluates the transformed Chebyshev series g and its gradient at a point x.

WaveGreen2D.Chebyshev.gradient — Method

gradient(f::AbstractChebyshevSeries{T, 1}, x::T) where T -> T, T

Simpler function for evaluating a one-dimensional Chebyshev series f and its gradient at a point x, where x is of type T.

WaveGreen2D.Chebyshev.gradient_clenshaw — Method

gradient_clenshaw(a::Array{T, N}, x::SVector{N, T}) where {T, N} -> T, SVector{N, T}

Implements the Clenshaw algorithm to evaluate the N-th dimensional Chebyshev series with coefficients a and its gradient at a normalized point x in $[-1, 1]^N$.

WaveGreen2D.Chebyshev.gradient_clenshaw — Method

gradient_clenshaw(a::Array{T, N}, x::T) where {T, N} -> Array{T, N-1}, Array{T, N-1}

Implements the Clenshaw algorithm to evaluate the N-th dimensional Chebyshev series with coefficients a and its gradient at a normalized value x of its N-th dimension.

WaveGreen2D.Chebyshev.hessian — Method

hessian(
    f::AbstractChebyshevSeries{T, N}, x::AbstractVector{T}
) where {T, N} -> T, SVector{N, T}, SMatrix{N, N, T}

Simpler function for evaluating a Chebyshev series f, its gradient and hessian at a point x, where x is of any subtype of an AbstractVector{T}.

WaveGreen2D.Chebyshev.hessian — Method

hessian(
    h::ChebyshevCluster{T, N}, x::SVector{N, T}
) where {T, N} -> T, SVector{N, T}, SMatrix{N, N, T}

Evaluates the Chebyshev cluster h, its gradient and hessian at a point x.

WaveGreen2D.Chebyshev.hessian — Method

hessian(
    f::ChebyshevSeries{T, N}, x::SVector{N, T}
) where {T, N} -> T, SVector{N, T}, SMatrix{N, N, T}

Evaluates the Chebyshev series f, its gradient and hessian at a point x.

WaveGreen2D.Chebyshev.hessian — Method

hessian(
    g::TransformedChebyshevSeries{T, N}, x::SVector{N, T}
) where {T, N} -> T, SVector{N, T}, SMatrix{N, N, T}

Evaluates the transformed Chebyshev series g, its gradient and hessian at a point x.

WaveGreen2D.Chebyshev.hessian — Method

hessian(f::AbstractChebyshevSeries{T, 1}, x::T) where T -> T, T, T

Simpler function for evaluating a one-dimensional Chebyshev series f, its gradient and hessian at a point x, where x is of type T.

WaveGreen2D.Chebyshev.hessian_clenshaw — Method

hessian_clenshaw(
    a::Array{T, N}, x::SVector{N, T}
) where {T, N} -> T, SVector{N, T}, SMatrix{N, N, T}

Implements the Clenshaw algorithm to evaluate the N-th dimensional Chebyshev series with coefficients a, its gradient and hessian at a normalized point x in $[-1, 1]^N$.

WaveGreen2D.Chebyshev.hessian_clenshaw — Method

hessian_clenshaw(
    a::Array{T, N}, x::T
) where {T, N} -> Array{T, N-1}, Array{T, N-1}, Array{T, N-1}

Implements the Clenshaw algorithm to evaluate the N-th dimensional Chebyshev series with coefficients a, its gradient and hessian at a normalized value x of its N-th dimension.

WaveGreen2D.Chebyshev.normalize — Method

normalize(f::ChebyshevSeries{T, N}, x::SVector{N, T}) where {T, N} -> SVector{N, T}

Converts a point x to its normalized coordinates in $[-1, 1]^N$.

WaveGreen2D.Chebyshev.order — Method

order(f::ChebyshevSeries{T,1}) where T -> Int

WaveGreen2D.Chebyshev.order — Method

order(f::ChebyshevSeries{T,N}) where {T,N} -> NTuple{N, Int}

Chebyshev series order, which is equal to the number of coefficients along each dimension minus one.

WaveGreen2D.Chebyshev.symmatrix — Method

symmatrix(u::SVector{K, T}, ::Val{N}) where {T, N, K} -> SMatrix{N, N, T}

Converts a vector of K values representing the upper triangular matrix of order N, stored in column-major order, into a symmetric matrix of order N. It's necessary that K = N*(N+1)÷2.

WaveGreen2D.Chebyshev.validate_transformation — Method

validate_transformation(::Type{T}, N::Int, u, ∇u, Hu) where T

Validates the input and return types of u, ∇u and Hu.